使用BigDecimal来处理浮点运算

+++ src/main/java/com/rhenium/meethere/domain/Booking.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.apache.tomcat.jni.Local;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -32,7 +32,7 @@ public class Booking implements Serializable {
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     private LocalDateTime endTime;
 
-    private Integer priceSum;
+    private BigDecimal priceSum;
     private Boolean paid;
 
     /** For SQL **/
